requirements.txt
google-genai
python-dotenv
streamlit
pydantic
pandas
pytest
pypdf
